Make the pytest configuration reusable
